F: Updated MapGenerator further
Updated MapGenerator with the methods needed to set up connections between generated grids, and ensuring that those connections will not cross existing paths as they are being connected, as well as not going out of bounds of the map's size. Some debugging code is found  at the "generate_map" function to test these feautures.