H: Created Map Room Scene + Debugging Code
Following up with the map generator, the actual scene for the Map was made, complete with animations for the selection and highlight for the individual nodes. A dictionary, ICONS, was made to preload the respective sprites and their scale Vector2's to be compatible with the MapRoom Area2D scene.