R: Created Run Scene and Template Room Scenes
Run scene is operational, and functionality to start a new run to generate the map, and switch between relevant scenes upon the player's selection on the map scene. Template versions of the battle, shop, treasure room, battle rewards, and campfire scenes were made to ensure that the run.gd scene was operational.