S: Started Effects Framework, Tooltip, Planning Ability GUI
Made a few default effects for abilities (damage, armor, mana-burn), and some sample abilities utilizing the effects system with their resources. Designed a tooltip template to show effects of abilities (and maybe statuses in the future) if the player hover's over the ability button for a short duration. More work is needed to design the ability hotbar and targeting functionality.