2C: Impletmented Enemy AI for Crab Enemy, Turn Based Combat Operational
Enemies now derive from the EnemyStats script, which includes an exporyt for their individual AI scenes. Within the AI scenes, a tree of nodes starting with an EnemyActionPicker, and the enemy's list of abilities/moves are shown. Each individual node describing their abilities has a seperate script which instantiates an effect, sets an "amount" for that effect, references their ability as a new ability set in the ediotr, references the enemy itself to extract their Attributes to calculate scaling modifiers, and executes the effect with some added tweens for presentability.

There is now a simple Win/Lose condition operational in the BattleScene. To win, defeat all enemies within the enemy container. The lose condition is earned by losing all health and dying in combat. Debug statements for damage calculations for enemy actions, as well as the Win/Lose conditions, have been implemented for clarity of function.