2D: Implemented Enemy Intent System; Updates Values based on Data
Implemented a simple intent system including a "Number" and a subsequent "Icon" for enemy actions. Added methods to calculate expected incoming damage and armor gains from enemy sources, which is connected to their stats and Scalings, meaning that it should change depending on their attributes changing mid-combat. A global signal may be needed to recalculate the enemy intent text to keep up with attribute debuffs which may be applied to enemy targets. This will be considered when status effects are added at a later date.