2F: Two Shaders, Hit Animations, Player and Enemy SFX, Music and Sound Player Audio Buses
Created mini animations for when the player or enemy gets hit by an attack (one that would normally be triggered by the [take_damage] function). When hit, the target will flash white, if they have remaining armor after the attack, or red, and have a "Shake" effect occur to indicate a hit has been recieved. Music and SFX AudioBuses have been created, allowing for the playing and updating of music and sound effects in the game through two global/autoload scenes.