2G: Added New Bat Enemy, Organizing Templates for Future Enemy Attacks, Effects, and Abilities
A new enemy, "Bat", has been added to the Enemies folder in "Resources". This enemy is the first to have multiple instances of their action occur, this time in the form of two-attacks in one turn. The enemy intent updating function of Enemy.gd was modified to consider showing the number of effect instances that will occur from their action, showing a "x {effect_count}" next to the calculated damage number in the intent UI.

The configuration of templates has begun for the future development of the project. So far, a simple EnemyAttackActionTemplate and EnemyDefenseActionTemplate have been added to provide a simple starting point to new enemy actions, saving tons of time.