2K: Implemented Basic Spellbook into Combat and "Overworld" menus using dedicated button
The spellbook now properly references the "known_abilities" of the player and can be seen by pressing the dedicated "SpellBook" button in either the map menu or the battle menu. Upon clicking on any of the abilities listed, the player can get a description for the ability to read about its effects.