2O: Updating Inventory and Item System, new JSON file in the works for Items
The items and inventory system are displaying properly, some test cases have been made in their respective scripts to emulate a random assortment of custom items made with sprites (equipment models, potions, etc.)

TODO: Configure a stacking and equiping mechanic for the items, and figure out how to alter the stats and attributes of the player equiping items. Furthermore, configure consumable items to "cast" a self or targeted effect to heal, damage, and eventually status effects.