2P: Adjustments to Topbar and PlayerUI (CombatUI)
"GUICornerButtons" and their functionality has been removed in exchange for having the top-bar's buttons providing the functionality of showing the player's map, abilities, equipment/inventory screens. Further implementation is needed to configure the player with their inventory per run, instead of being randomly generated as per testing purposes, yet the displaying of the inventory screen is working as intended for now.