2Q: Further Updates to Inventory/Items
Additional functionalities were added to the inventory system in progress. A new global(autoload) titled "PlayerInventory" has been added to allow for all scenes within the run to access the dictionary of items that the player will hold in their individual runs.

TODO: With this new system, determine if its possible to store a dictionary of items per class, this could potentially serve as "starting items" for players!