2R: Added Right Click Functions to the Inventory System
When holding an item, added right click functions to split the held item off by one, allowing for individual placement of items in their slots, and adding one of the held item to a slot containing the same item data.

TODO: Hook up a mini-menu on right-clicking without an item in hand to enable a "USE ITEM" function from the inventory screen.