2S: Inventory Updates at Runtime
Added some code to allow for the updating of the inventory's visuals when items are added during run time. However, the updating of visuals when items are removed during run time is still in progress, as "PlayerInventory.remove_item()" does remove the item from the PlayerInventory inventory dictionary, yet leaves the icon of the item in the inventory visual scene.