2T: Another Inventory and Item Overhaul
The Inventory and Item System have been overhauled again, this time with the items being stored as copies of a ItemData resource file, similiarly to how the abilities in the game are made. With this system, I can adjust their visual and gameplay attributes just like how I would for abilities, enemies, and the player's classes.

Furthermore, the InventoryUI scene has been updated to include two more panels on the left, an EquipmentPanel and DescriptionPanel. The equipment panel will be worked on in the future when equipment items have been configured ultilizing this system. The Description panel has been configured to overlay a focused item's visual attributes, such as its name, rarity color, item type, description, consumption or equipping effects (a work in progress, but a placeholder label for now), and the gold value of the item.