2Y: GoldRewards and AbilityRewards in Progress
Gold rewards and ability rewards for the BattleRewards scene are in progress, with GoldRewards being functional and increasing and setting the run_scene's gold variable properly. The AbilityRewards scene is being made to be reusable across other scenes of the game, not just combat but also the shop scene and perhaps random event screens too!