2Z: BattleRewards Functional, Test Code Implemented, Run Scene Working with Systems
The BattleRewards scene and scripts are all functioning as intended. Abilities now have a Rarity enum to help determine what abilities are rolled at the time of rewards. In the future.

TODO: Create a way to prevent the crashing when there are no suitable cards for an ability of a chosen rarity. Perhaps instantiate a default treasure card in its place?