3J: Status Effects Pt 1
Added templates for the majority of the status effects I will be implementing in the game. However, the buff components of status effects like "Empowered" are not functional yet, as they do not increase the stats of the status effect's owner. This will be done in the next commit with Modifiers.