3M: Refactoring of Tooltips and Fixed Bug with AbilitySlot "targets" Array Causing Crashing
Major refactoring of the tooltips for all abilities currently in the game (templates adjusted as well). Furthermore, a bug has been resolved with the "targets" array of the AbilitySlot script, where the "request_tooltip" function would be called on an enemy that has already been queue_freed, resulting in a crash.