3N: Finished Attribute Status Effect Implementation, Removal of Unnecessary Files, Small Bug Fixes
All of the attribute status effects are properly implemented and working. New icons for the status effects (existing and new) have been added to the assets folder. Bug fixes have been made to prevent going to 0 health when using an ability with a health cost that would bring you to 0. Unused scenes and scripts have been removed.