3P: Implemented "Chill" Effect using DMG_DEALT Modifiers; updates to damage calculations by enemies and player
The Chill Status effect has been implemented. Although there is a slight issue in the calculation of the damage since it doesnt take into consideration the calculation of the "p_s_mod" and "s_s_mod" of the attacks. Adjustments to include the "DMG_DEALT" modifiers to player abilities and existing enemy attack actions have been implemented.